about to move away from jacobian. It won't work for us because the core out also depends on mem_in. And core won't work with jacobian, so we can't include it in the calculation to find the grad of mem_in.
Instead we'll just slowly and wastefully loop through each value in memory
one at a time to calculate the grad of memory output wrt mem_in and
params.

Maybe we could move to a different library if this ends up as
crap as it seems. But if we do that, how can we attach ourselves to
existing models that use pytorch?

So then, the only other choice is to directly mod pytorch which seems
pretty bad. We will see.